+++ func/formulaProfile.php
@@ -1,5 +1,6 @@
 <?php 
 if (!defined('pvault_panel')){ die('Not Found');}
+require_once(__ROOT__.'/func/countElement.php');
 
 function formulaProfile($conn, $profile, $sex){
 
@@ -14,6 +15,7 @@ function formulaProfile($conn, $profile, $sex){
        <tr>
          <th width="20%">Formula Name</th>
          <th width="20%">Product Name</th>
+         <th width="20%">Ingredients</th>
          <th width="10%">Class</th>
          <th width="20%">Created</th>
          <th width="20%">Actions</th>
@@ -27,6 +29,7 @@ function formulaProfile($conn, $profile, $sex){
 		<tr>
 		<td align="center"><?php if($formula['isProtected']){?><a class="fas fa-lock"><?php } ?></a>&nbsp;<a href="?do=Formula&id=<?=$meta['id']?>"><?=$formula['name']?></a></td>
 		<td align="center"><a href="pages/getFormMeta.php?id=<?php echo $meta['id']; ?>" class="popup-link"><?=$meta['product_name']?></a></td>
+        <td align="center"><?php echo countElement("formulas WHERE fid = '".$meta['fid']."'",$conn);?></td>
 		<td align="center"><?php echo ucfirst($meta['catClass']);?></td>
 		<td align="center"><?php echo $meta['created'];?></td>
 		<td align="center"><a href="pages/getFormMeta.php?id=<?php echo $meta['id']; ?>" class="fas fa-comment-dots popup-link"></a> &nbsp; <a href="javascript:addTODO('<?php echo $formula['fid']; ?>')" class="fas fa-list" rel="tipsy" title="Add <?php echo $formula['name']; ?> to the make list" ></a>  &nbsp; <a href="javascript:cloneMe('<?php echo $formula['fid']; ?>')" class="fas fa-copy" rel="tipsy" title="Clone <?php echo $formula['name']; ?>"></a>  &nbsp; <a href="javascript:deleteMe('<?php echo $formula['fid']; ?>')" onclick="return confirm('Delete <?php echo $formula['name']; ?> Formula?')" class="fas fa-trash" rel="tipsy" title="Delete <?php echo $formula['name']; ?>"></a></td>