+++ pages/ingredients.php
@@ -378,4 +378,19 @@ function importING(name) {
 	  });
 };
 
+function setView(view) {
+	$.ajax({ 
+    url: 'pages/update_settings.php', 
+	type: 'get',
+    data: {
+		ingView: view,
+		},
+	dataType: 'html',
+    success: function (data) {
+		location.reload();
+    }
+  });
+};
+
+
 </script>