+++ pages/pvonline.php
@@ -5,6 +5,7 @@ require_once(__ROOT__.'/inc/config.php');
 require_once(__ROOT__.'/inc/opendb.php');
 require_once(__ROOT__.'/inc/settings.php');
 require_once(__ROOT__.'/inc/product.php');
+require_once(__ROOT__.'/func/pvOnline.php');
 
 
 if($_GET['action'] == 'import' && $_GET['items']){
@@ -51,7 +52,7 @@ if($_GET['action'] == 'import' && $_GET['items']){
 }
 
 if($_GET['action'] == 'upload' && $_GET['items'] == 'ingredients'){
-	//Do all the ingredients
+	//Upload all the ingredients
 	$ingQ = mysqli_query($conn, "SELECT * FROM ingredients WHERE isPrivate = '0'");
 	$i = 0;
 	while($ing = mysqli_fetch_assoc($ingQ)){
@@ -59,29 +60,49 @@ if($_GET['action'] == 'upload' && $_GET['items'] == 'ingredients'){
 		if($_GET['excludeNotes'] == 'true'){
 			unset($ing['notes']);			
 		}
-		$ar = array_filter($ing);
-		
-		$url = http_build_query($ar);
-		$jAPI = $pvOnlineAPI.'?username='.$pv_online['email'].'&password='.$pv_online['password'].'&do=add&kind=ingredient&'.$url;
+		$data = array_filter($ing);
+		$data["username"] = $pv_online['email'];
+		$data["password"] = $pv_online['password'];
+		$data["do"] = "add";
+		$data["kind"] = "ingredient";
+	
+		$up_req.= pvUploadData($pvOnlineAPI, $data);
 		$i++;
-		$up_req = file_get_contents($jAPI,true);
 	}
 	
-	//Do all the allergens
+	//Upload all the allergens
 	$algQ = mysqli_query($conn, "SELECT * FROM allergens");
+	$a = 0;
 	while($alg = mysqli_fetch_assoc($algQ)){
 		unset($alg['id']);
-		$ar = array_filter($alg);
+		$data = array_filter($alg);
+		$data["username"] = $pv_online['email'];
+		$data["password"] = $pv_online['password'];
+		$data["do"] = "add";
+		$data["kind"] = "allergen";
 		
-		$url = http_build_query($alg);
-		$jAPI = $pvOnlineAPI.'?username='.$pv_online['email'].'&password='.$pv_online['password'].'&do=add&kind=allergen&'.$url;
-		$up_req.= file_get_contents($jAPI,true);
+		$up_req.= pvUploadData($pvOnlineAPI, $data);
+		$a++;
 	}
 	
+	//Upload all the categories
+	$alC = mysqli_query($conn, "SELECT id, name, notes, image, colorKey FROM ingCategory");
+	$c = 0;
+	while($cat = mysqli_fetch_assoc($alC)){
+		$data = array_filter($cat);
+		$data["username"] = $pv_online['email'];
+		$data["password"] = $pv_online['password'];
+		$data["do"] = "add";
+		$data["kind"] = "category";
+	
+		$up_req.= pvUploadData($pvOnlineAPI, $data);
+		$c++;
+	}
 	if($up_req){
-		echo  '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>'.$i.' ingredients uploaded!</div>';
+		echo  '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>'.$i.'</strong> ingredients, <strong>'.$a.'</strong> allergens and <strong>'.$c.'</strong> categories uploaded!</div>';
 	}
 
 	return;
 }
+
 ?>